requirements.txt
numpy
python-dotenv
gradio
tqdm
Pillow